coding master  changed param from long to float

+++ src/main/java/com/operation/calculator/model/Calculator.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @AllArgsConstructor @Getter
 public class Calculator {
 
-	private final long result;
+	private final float result;
 
 }
